% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veto-package.R
\docType{package}
\name{veto-package}
\alias{veto-package}
\alias{_PACKAGE}
\alias{veto_package}
\title{veto: Simple Error Messages for Conditions}
\description{
Simplified functions intended to check function parameters to
    ensure they meet specific requirements before continuing; otherwise vetoed
    and stopped with an error.
}
\author{
\strong{Maintainer}: Jordan Mark Barbone \email{jmbarbone@gmail.com} (\href{https://orcid.org/0000-0001-9788-3628}{ORCID}) [copyright holder]

}
\keyword{internal}
