% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veto.R
\name{veto}
\alias{veto}
\title{Veto}
\usage{
veto(cond, msg)
}
\arguments{
\item{cond}{A condition, \code{TRUE} or \code{FALSE}}

\item{msg}{A}
}
\value{
None, called for side effects
}
\description{
Veto something
}
