% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{veto_unless}
\alias{veto_unless}
\title{Veto unless}
\usage{
veto_unless(parameter, condition, message)
}
\arguments{
\item{parameter}{A parameter to check}

\item{condition}{A condition that evaluates to either \code{TRUE} or \code{FALSE}; if
\code{FALSE} a \code{veto} will be triggered}

\item{message}{The \code{veto} message to display}
}
\value{
None, called for side effects
}
\description{
A veto'ing template
}
\examples{
foo <- function(x) {
  veto_unless(x, is.character(x), "must be a character")
  sprintf("x = '\%s' is a character", x)
}

foo("letters")
try(foo(1:10))
}
