% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{veto_wrappers}
\alias{veto_wrappers}
\alias{veto_character}
\alias{veto_integer}
\alias{veto_double}
\alias{veto_numeric}
\alias{veto_date}
\alias{veto_datetime}
\alias{veto_class}
\alias{veto_length}
\alias{veto_nchar}
\alias{veto_true_false}
\title{Veto wrappers}
\usage{
veto_character(x, length = NULL)

veto_integer(x, length = NULL)

veto_double(x, length = NULL)

veto_numeric(x, length = NULL)

veto_date(x, length = NULL)

veto_datetime(x, length = NULL)

veto_class(x, class, length = NULL)

veto_length(x, length = 1L)

veto_nchar(x, n = 1)

veto_true_false(x)
}
\arguments{
\item{x}{A parameter}

\item{length}{The expected length}

\item{class}{Character vector of the required class}

\item{n}{The number of}
}
\value{
None, called for side effects
}
\description{
Veto wrappers
}
\examples{
foo <- function(a, b) {
  veto_integer(a)
  veto_double(b, 2L)
  a * prod(b)
}

foo(1L, c(0.5, 1))
try(foo(1L, 0))
}
